+++ database/factories/TestUserFactory.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace DutchCodingCompany\FilamentDeveloperLogins\Database\Factories;
+namespace Patriciomartins\FilamentDeveloperLogins\Database\Factories;
 
-use DutchCodingCompany\FilamentDeveloperLogins\Tests\Fixtures\TestUser;
+use Patriciomartins\FilamentDeveloperLogins\Tests\Fixtures\TestUser;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;