+++ tests/Fixtures/create_users_table.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DutchCodingCompany\FilamentDeveloperLogins\Tests\Fixtures;
+namespace Patriciomartins\FilamentDeveloperLogins\Tests\Fixtures;
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;